+++ product_harmonized_system_per_country/models/__init__.py
@@ -6,3 +6,4 @@ from . import account_move
 from . import hs_code
 from . import product_category
 from . import product_product
+from . import product_template
